v1.1.0
i customized more of the fonts, and i updated the email so it actually works. i also made the load-up of the portfolio look cooler and updated the project filters so that the buttons are in alphabetical order and more projects are designated in their filter options more accurately. now the website is quite fully functional, but i have yet to update the projects that are uploaded with more videos, and upload more past projects...